+++ src/collective/feedaggregator/browser/__init__.py
@@ -2,6 +2,7 @@
 from collective.feedaggregator.config import TTL
 from plone import api
 from plone.memoize import ram
+from Products.CMFPlone.PloneBatch import Batch
 from Products.Five.browser import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 
@@ -24,6 +25,12 @@ class ListingView(BrowserView):
 
     index = ViewPageTemplateFile('feedaggregator.pt')
 
+    def __init__(self, context, request):
+        super(ListingView, self).__init__(context, request)
+        self.b_size = context.item_count
+        b_start = getattr(request, 'b_start', 0)
+        self.b_start = int(b_start)
+
     def __call__(self):
         return self.index()
 
@@ -31,6 +38,10 @@ class ListingView(BrowserView):
     @ram.cache(_feedaggregator_cachekey)
     def results(self):
         return self.context.results()
+
+    def batch(self):
+        return Batch(
+            self.results, size=self.b_size, start=self.b_start, orphan=1)
 
     @property
     def show_byline(self):